services/vanna/requirements.txt
fastapi
uvicorn
vanna           # 1. Base Vanna package
groq            # 2. Groq ka asli package (jo 'vanna[groq]' ko install karna tha)
psycopg2-binary # 3. Postgres ka package
python-dotenv
gunicorn